move stuff out of shared that is not shared.
